wait for all batches, create pubsub service
